refactor to use index.tsx
